Switch places between the client receiver commands module and the client sender commands module
