+++ src/Versions/V2_2_1/Client/Receiver/Commands.php
@@ -5,15 +5,43 @@ declare(strict_types=1);
 namespace Chargemap\OCPI\Versions\V2_2_1\Client\Receiver;
 
 use Chargemap\OCPI\Common\Client\Modules\AbstractFeatures;
-use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\PostCommandResultRequest;
-use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\PostCommandResultResponse;
-use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\PostCommandResultService;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\CancelReservation\CancelReservationRequest;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\CancelReservation\CancelReservationService;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\CommandResponseResponse;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\ReserveNow\ReserveNowRequest;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\ReserveNow\ReserveNowService;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\StartSession\StartSessionRequest;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\StartSession\StartSessionService;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\StopSession\StopSessionRequest;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\StopSession\StopSessionService;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\UnlockConnector\UnlockConnectorRequest;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\UnlockConnector\UnlockConnectorService;
 
 class Commands extends AbstractFeatures
 {
-    
-    public function post(PostCommandResultRequest $request): PostCommandResultResponse
+
+    public function cancelReservation(CancelReservationRequest $request): CommandResponseResponse
     {
-        return (new PostCommandResultService($this->ocpiConfiguration))->handle($request);
+        return (new CancelReservationService($this->ocpiConfiguration))->handle($request);
+    }
+
+    public function reserveNow(ReserveNowRequest $request): CommandResponseResponse
+    {
+        return (new ReserveNowService($this->ocpiConfiguration))->handle($request);
+    }
+
+    public function startSession(StartSessionRequest $request): CommandResponseResponse
+    {
+        return (new StartSessionService($this->ocpiConfiguration))->handle($request);
+    }
+
+    public function stopSession(StopSessionRequest $request): CommandResponseResponse
+    {
+        return (new StopSessionService($this->ocpiConfiguration))->handle($request);
+    }
+
+    public function unlockConnector(UnlockConnectorRequest $request): CommandResponseResponse
+    {
+        return (new UnlockConnectorService($this->ocpiConfiguration))->handle($request);
     }
 }
