+++ src/Versions/V2_2_1/Client/Receiver/Commands/Post/CommandResponseResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Chargemap\OCPI\Versions\V2_2_1\Client\Sender\Commands\Post;
+namespace Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post;
 
 use Chargemap\OCPI\Common\Client\Modules\AbstractResponse;
 use Chargemap\OCPI\Versions\V2_2_1\Common\Factories\CommandResponseFactory;
@@ -20,7 +20,7 @@ class CommandResponseResponse extends AbstractResponse
 
     public static function fromResponseInterface(ResponseInterface $response): self
     {
-        $json = self::toJson($response, 'V2_2_1/Sender/Commands/commandResponsePostRequest.schema.json');
+        $json = self::toJson($response, 'V2_2_1/Receiver/Commands/commandResponsePostRequest.schema.json');
 
         return new self(CommandResponseFactory::fromJson($json->data));
     }
