+++ src/Versions/V2_2_1/Client/Receiver/Commands/Post/StopSession/StopSessionService.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Chargemap\OCPI\Versions\V2_2_1\Client\Sender\Commands\Post\UnlockConnector;
+namespace Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\StopSession;
 
 use Chargemap\OCPI\Common\Client\Modules\AbstractFeatures;
-use Chargemap\OCPI\Versions\V2_2_1\Client\Sender\Commands\Post\CommandResponseResponse;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\CommandResponseResponse;
 
-class UnlockConnectorService extends AbstractFeatures
+class StopSessionService extends AbstractFeatures
 {
     /**
-     * @param UnlockConnectorRequest $request
+     * @param StopSessionRequest $request
      * @return CommandResponseResponse
      * @throws \Chargemap\OCPI\Common\Client\OcpiEndpointNotFoundException
      * @throws \Psr\Http\Client\ClientExceptionInterface
      */
-    public function handle(UnlockConnectorRequest $request): CommandResponseResponse
+    public function handle(StopSessionRequest $request): CommandResponseResponse
     {
         $responseInterface = $this->sendRequest($request);
         return CommandResponseResponse::fromResponseInterface($responseInterface);
