+++ src/Versions/V2_2_1/Client/Receiver/Commands/Post/UnlockConnector/UnlockConnectorRequest.php
@@ -2,26 +2,26 @@
 
 declare(strict_types=1);
 
-namespace Chargemap\OCPI\Versions\V2_2_1\Client\Sender\Commands\Post\ReserveNow;
+namespace Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\UnlockConnector;
 
 use Chargemap\OCPI\Common\Client\Modules\AbstractRequest;
 use Chargemap\OCPI\Versions\V2_2_1\Client\VersionTrait;
 use Chargemap\OCPI\Versions\V2_2_1\Common\Models\ModuleId;
-use Chargemap\OCPI\Versions\V2_2_1\Common\Models\ReserveNow;
+use Chargemap\OCPI\Versions\V2_2_1\Common\Models\UnlockConnector;
 use Http\Discovery\Psr17FactoryDiscovery;
 use Psr\Http\Message\ServerRequestFactoryInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamFactoryInterface;
 
-class ReserveNowRequest extends AbstractRequest
+class UnlockConnectorRequest extends AbstractRequest
 {
     use VersionTrait;
 
-    private ?ReserveNow $stopSession;
+    private ?UnlockConnector $unlockConnector;
 
-    public function __construct(ReserveNow $stopSession)
+    public function __construct(UnlockConnector $unlockConnector)
     {
-        $this->stopSession = $stopSession;
+        $this->unlockConnector = $unlockConnector;
     }
 
     public function getModule(): ModuleId
@@ -35,11 +35,11 @@ class ReserveNowRequest extends AbstractRequest
             $streamFactory = Psr17FactoryDiscovery::findStreamFactory();
         }
 
-        $request = $serverRequestFactory->createServerRequest('POST', '/RESERVE_NOW');
+        $request = $serverRequestFactory->createServerRequest('POST', '/UNLOCK_CONNECTOR');
 
-        if ($this->stopSession) {
+        if ($this->unlockConnector) {
             $request = $request->withHeader('Content-Type', 'application/json')
-                ->withBody($streamFactory->createStream(json_encode($this->stopSession)));
+                ->withBody($streamFactory->createStream(json_encode($this->unlockConnector)));
         }
 
         return $request;
