+++ src/Versions/V2_2_1/Client/Receiver/Commands/Post/UnlockConnector/UnlockConnectorService.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Chargemap\OCPI\Versions\V2_2_1\Client\Sender\Commands\Post\StopSession;
+namespace Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\UnlockConnector;
 
 use Chargemap\OCPI\Common\Client\Modules\AbstractFeatures;
-use Chargemap\OCPI\Versions\V2_2_1\Client\Sender\Commands\Post\CommandResponseResponse;
+use Chargemap\OCPI\Versions\V2_2_1\Client\Receiver\Commands\Post\CommandResponseResponse;
 
-class StopSessionService extends AbstractFeatures
+class UnlockConnectorService extends AbstractFeatures
 {
     /**
-     * @param StopSessionRequest $request
+     * @param UnlockConnectorRequest $request
      * @return CommandResponseResponse
      * @throws \Chargemap\OCPI\Common\Client\OcpiEndpointNotFoundException
      * @throws \Psr\Http\Client\ClientExceptionInterface
      */
-    public function handle(StopSessionRequest $request): CommandResponseResponse
+    public function handle(UnlockConnectorRequest $request): CommandResponseResponse
     {
         $responseInterface = $this->sendRequest($request);
         return CommandResponseResponse::fromResponseInterface($responseInterface);
